Support for draft 3
