+++ tests/unit/s2n_npn_extension_test.c
@@ -67,7 +67,7 @@ int main(int argc, char **argv)
         EXPECT_TRUE(s2n_client_alpn_extension.should_send(client_conn));
 
         /*
-         *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-04#section-3
+         *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-03#section-3
          *= type=test
          *# For the same reasons, after a handshake has been performed for a
          *# given connection, renegotiations on the same connection MUST NOT
@@ -201,7 +201,7 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_stuffer_write_bytes(&extension, protocol, sizeof(protocol)));
 
             /*
-             *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-04#section-4
+             *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-03#section-4
              *= type=test
              *# In the event that the client doesn't support any of server's protocols, or
              *# the server doesn't advertise any, it SHOULD select the first protocol
@@ -225,7 +225,7 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&extension, 0));
 
             /*
-             *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-04#section-4
+             *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-03#section-4
              *= type=test
              *# In the event that the client doesn't support any of server's protocols, or
              *# the server doesn't advertise any, it SHOULD select the first protocol
@@ -264,156 +264,6 @@ int main(int argc, char **argv)
 
             /* Client's second protocol is selected because the server prefers it over client's first protocol */
             EXPECT_BYTEARRAY_EQUAL(s2n_get_application_protocol(client_conn), protocols[1], strlen(protocols[1]));
-        }
-    }
-
-    /* Tests for the NPN extension on the Encrypted Extensions handshake message */
-    {
-        /* s2n_npn_encrypted_should_send */
-        {
-            /* No connection */
-            EXPECT_FALSE(s2n_npn_encrypted_extension.should_send(NULL));
-
-            /* No chosen application protocol */
-            DEFER_CLEANUP(struct s2n_connection *client_conn = s2n_connection_new(S2N_CLIENT), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(client_conn);
-            EXPECT_FALSE(s2n_npn_encrypted_extension.should_send(client_conn));
-
-            /* Should send if protocol has been negotiated */
-            uint8_t first_protocol_len = strlen(protocols[0]);
-            EXPECT_MEMCPY_SUCCESS(client_conn->application_protocol, protocols[0], first_protocol_len + 1);
-            EXPECT_TRUE(s2n_npn_encrypted_extension.should_send(client_conn));
-        }
-
-        /* s2n_npn_encrypted_extension_send */
-        {
-            DEFER_CLEANUP(struct s2n_connection *client_conn = s2n_connection_new(S2N_CLIENT), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(client_conn);
-
-            uint8_t first_protocol_len = strlen(protocols[0]);
-            EXPECT_MEMCPY_SUCCESS(client_conn->application_protocol, protocols[0], first_protocol_len + 1);
-
-            DEFER_CLEANUP(struct s2n_stuffer out = { 0 }, s2n_stuffer_free);
-            EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&out, 0));
-            EXPECT_SUCCESS(s2n_npn_encrypted_extension.send(client_conn, &out));
-
-            uint8_t protocol_len = 0;
-            EXPECT_SUCCESS(s2n_stuffer_read_uint8(&out, &protocol_len));
-
-            uint8_t protocol[UINT8_MAX] = { 0 };
-            EXPECT_SUCCESS(s2n_stuffer_read_bytes(&out, protocol, protocol_len));
-            EXPECT_BYTEARRAY_EQUAL(protocol, protocols[0], protocol_len);
-
-            uint8_t padding_len = 0;
-            EXPECT_SUCCESS(s2n_stuffer_read_uint8(&out, &padding_len));
-            EXPECT_TRUE(padding_len > 0);
-            
-            for(size_t i = 0; i < padding_len; i++) {
-                uint8_t byte = UINT8_MAX;
-                EXPECT_SUCCESS(s2n_stuffer_read_uint8(&out, &byte));
-                EXPECT_EQUAL(byte, 0);
-            }
-            EXPECT_EQUAL(s2n_stuffer_data_available(&out), 0);
-        }
-
-        /* NPN Encrypted Extension recv can read NPN Encrypted Extension send */
-        {
-            DEFER_CLEANUP(struct s2n_connection *client_conn = s2n_connection_new(S2N_CLIENT), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(client_conn);
-
-            DEFER_CLEANUP(struct s2n_connection *server_conn = s2n_connection_new(S2N_SERVER), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(server_conn);
-
-            uint8_t first_protocol_len = strlen(protocols[0]);
-            EXPECT_MEMCPY_SUCCESS(client_conn->application_protocol, protocols[0], first_protocol_len + 1);
-
-            DEFER_CLEANUP(struct s2n_stuffer out = { 0 }, s2n_stuffer_free);
-            EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&out, 0));
-            EXPECT_SUCCESS(s2n_npn_encrypted_extension.send(client_conn, &out));
-            EXPECT_SUCCESS(s2n_npn_encrypted_extension.recv(server_conn, &out));
-
-            EXPECT_NOT_NULL(s2n_get_application_protocol(server_conn));
-            EXPECT_BYTEARRAY_EQUAL(s2n_get_application_protocol(server_conn), protocols[0], strlen(protocols[0]));
-        }
-
-        /* NPN Encrypted Extension recv can read empty extension */
-        {
-            DEFER_CLEANUP(struct s2n_connection *server_conn = s2n_connection_new(S2N_SERVER), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(server_conn);
-
-            DEFER_CLEANUP(struct s2n_stuffer out = { 0 }, s2n_stuffer_free);
-            EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&out, 0));
-
-            /* Zero-length protocol */
-            EXPECT_SUCCESS(s2n_stuffer_write_uint8(&out, 0));
-            uint8_t padding_len = 0;
-            EXPECT_OK(s2n_calculate_padding(0, &padding_len));
-            EXPECT_SUCCESS(s2n_stuffer_write_uint8(&out, padding_len));
-            for(size_t i = 0; i < padding_len; i++) {
-                EXPECT_SUCCESS(s2n_stuffer_write_uint8(&out, 0));
-            }
-
-            EXPECT_SUCCESS(s2n_npn_encrypted_extension.recv(server_conn, &out));
-            EXPECT_NULL(s2n_get_application_protocol(server_conn));
-        }
-
-        /* NPN Encrypted Extension recv errors on malformed extension */
-        {
-            DEFER_CLEANUP(struct s2n_connection *server_conn = s2n_connection_new(S2N_SERVER), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(server_conn);
-
-            uint8_t wire_bytes[] = {
-                /* Incorrect length of extension */
-                0x10,
-            };
-
-            DEFER_CLEANUP(struct s2n_stuffer out = { 0 }, s2n_stuffer_free);
-            EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&out, 0));
-            EXPECT_SUCCESS(s2n_stuffer_write_bytes(&out, wire_bytes, sizeof(wire_bytes)));
-            EXPECT_FAILURE_WITH_ERRNO(s2n_npn_encrypted_extension.recv(server_conn, &out), S2N_ERR_NULL);
-        }
-
-        /* NPN Encrypted Extension recv errors on malformed padding */
-        {
-            DEFER_CLEANUP(struct s2n_connection *server_conn = s2n_connection_new(S2N_SERVER), s2n_connection_ptr_free);
-            EXPECT_NOT_NULL(server_conn);
-
-            uint8_t wire_bytes[] = {
-                /* Zero-length protocol */
-                0x00, 0x00,
-                /* Padding character is not zero */
-                0x01, 0xFF,
-            };
-
-            DEFER_CLEANUP(struct s2n_stuffer out = { 0 }, s2n_stuffer_free);
-            EXPECT_SUCCESS(s2n_stuffer_growable_alloc(&out, 0));
-            EXPECT_SUCCESS(s2n_stuffer_write_bytes(&out, wire_bytes, sizeof(wire_bytes)));
-            EXPECT_FAILURE_WITH_ERRNO(s2n_npn_encrypted_extension.recv(server_conn, &out), S2N_ERR_SAFETY);
-        }
-
-        /*
-         *= https://datatracker.ietf.org/doc/id/draft-agl-tls-nextprotoneg-04#section-3
-         *= type=test
-         *# The length of "padding" SHOULD be 32 - ((len(selected_protocol) + 2) % 32).
-         */
-        {
-            struct {
-                uint8_t protocol_len;
-                uint8_t expected_padding;
-            } test_cases[] = {
-                { .protocol_len = 0, .expected_padding = 30 },
-                { .protocol_len = 32, .expected_padding = 30 },
-                { .protocol_len = 17, .expected_padding = 13 },
-                { .protocol_len = 2, .expected_padding = 28 },
-                { .protocol_len = UINT8_MAX, .expected_padding = 31 },
-                { .protocol_len = 30, .expected_padding = 32 },
-            };
-
-            for (size_t i = 0; i < s2n_array_len(test_cases); i++) {
-                uint8_t output = 0;
-                EXPECT_OK(s2n_calculate_padding(test_cases[i].protocol_len, &output));
-                EXPECT_EQUAL(output, test_cases[i].expected_padding);
-            }
         }
     }
